+++ src/emulator/device/clock.rs
@@ -47,12 +47,11 @@ impl Device for Clock {
         0x1c6c8b36
     }
 
-    fn interrupt(&mut self, cpu: &mut Cpu)
-        -> InterruptResult {
+    fn interrupt(&mut self, cpu: &mut Cpu) -> Result<InterruptDelay> {
         let a = cpu.registers[Register::A];
         let b = cpu.registers[Register::B];
         match try!(Command::from_u16(a)
-                           .ok_or(InterruptError::InvalidCommand(a))) {
+                           .ok_or(ErrorKind::InvalidCommand(a))) {
             Command::SET_SPEED => self.speed = b,
             Command::GET_TICKS => {
                 cpu.registers[Register::C] = self.last_call as u16;