+++ src/emulator/device/mod.rs
@@ -15,18 +15,22 @@ pub enum TickResult {
 }
 
 pub type InterruptDelay = u16;
-#[derive(Debug)]
-pub enum InterruptError {
-    InvalidCommand(u16),
-}
-pub type InterruptResult = Result<InterruptDelay, InterruptError>;
+
+error_chain!(
+    errors {
+        InvalidCommand(cmd: u16) {
+            description("invalid hardware command")
+            display("invalid hardware command: {}", cmd)
+        }
+    }
+);
 
 pub trait Device: Debug {
     fn hardware_id(&self) -> u32;
     fn hardware_version(&self) -> u16;
     fn manufacturer(&self) -> u32;
 
-    fn interrupt(&mut self, &mut Cpu) -> InterruptResult;
+    fn interrupt(&mut self, &mut Cpu) -> Result<InterruptDelay>;
     fn tick(&mut self, &mut Cpu, current_tick: u64) -> TickResult;
 
     fn inspect(&self);