+++ src/lib.rs
@@ -1,5 +1,9 @@
 #![doc(html_root_url = "https://yamakaky.github.io/dcpu/")]
 
+#![recursion_limit = "1024"]
+#[macro_use]
+extern crate error_chain;
+
 #[macro_use]
 extern crate enum_primitive;
 #[macro_use]